% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_parameters.R
\name{attach_parameters}
\alias{attach_parameters}
\title{Attach pameters to the corresponsing trajectory}
\usage{
attach_parameters(trajectory, parameter_file, parameter_names = vector())
}
\arguments{
\item{trajectory}{The parsed trajectory}

\item{parameter_file}{The file with the stored parameters}

\item{parameter_names}{An optinal mapping that replaces integer identifiers with names for easy access}
}
\description{
Attach pameters to the corresponsing trajectory
}
