% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_species.R
\name{extract_species}
\alias{extract_species}
\title{Parses all unique entities from a json trajectory}
\usage{
extract_species(json_file)
}
\arguments{
\item{json_file}{The file with the trajectory data.}
}
\description{
Parses all unique entities from a json trajectory
}
