% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_collector.R
\name{map_collector}
\alias{map_collector}
\title{Maps the given entities to the given collectors.}
\usage{
map_collector(entites, criteria)
}
\arguments{
\item{entites}{The entities, which should be considered}

\item{criteria}{The criteria for mapping a entity to a collector}
}
\description{
Maps the given entities to the given collectors.
}
