% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_concentrations.R
\name{parse_concentrations}
\alias{parse_concentrations}
\title{Parses concentrations from a json trajectory.}
\usage{
parse_concentrations(json_file, select_nodes = vector(),
  select_sections = vector(), select_entities = vector())
}
\arguments{
\item{json_file}{The file with the trajectory data.}

\item{select_nodes}{A vector with nodes that will be considered, if no nodes are given all are considered}

\item{select_sections}{A vector with sections that will be considered, if no sections are given all are considered}

\item{select_entities}{A vector with entities that will be considered, if no entities are given all are considered}
}
\description{
Parses concentrations from a json trajectory.
}
