% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_logs.R
\name{parse_logs}
\alias{parse_logs}
\title{Parses all logs from an observation folder.}
\usage{
parse_logs(observation_folder, parameter_names = vector())
}
\arguments{
\item{parameter_names}{A reference for renaming parameter identifiers to parameter names}

\item{observation_directory}{The path to the observation directories (containing directories with a simulation each)}
}
\description{
Parses all logs from an observation folder.
}
