% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_vesicle_trajectories.R
\name{parse_vesicle_trajectories}
\alias{parse_vesicle_trajectories}
\title{Parses the states and positions of all vesicles over the course of the simulation}
\usage{
parse_vesicle_trajectories(json_file)
}
\arguments{
\item{json_file}{The file with the trajectory data.}
}
\description{
Parses the states and positions of all vesicles over the course of the simulation
}
