% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_experiment.R
\name{simplify_experiment}
\alias{simplify_experiment}
\title{Attach pameters to the corresponsing trajectory}
\usage{
simplify_experiment(experiment_directory, entity_mapping,
  parameter_names = vector())
}
\arguments{
\item{parameter_names}{An optinal mapping that replaces integer identifiers with names for easy access}

\item{trajectory}{The parsed trajectory}

\item{parameter_file}{The file with the stored parameters}
}
\description{
Attach pameters to the corresponsing trajectory
}
