% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_parameters.R
\name{simplify_parameters}
\alias{simplify_parameters}
\title{creates pairs of identifier and value from pameters}
\usage{
simplify_parameters(parameter_file, parameter_names = vector())
}
\arguments{
\item{parameter_file}{The file with the stored parameters}

\item{parameter_names}{An optinal mapping that replaces integer identifiers with names for easy access}
}
\description{
creates pairs of identifier and value from pameters
}
