% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_concentrations.R
\name{summarize_concentrations}
\alias{summarize_concentrations}
\title{Parses concentrations from a json trajectory.}
\usage{
summarize_concentrations(trajectory, entity_mapping)
}
\arguments{
\item{trajectory}{The parsed trajectory}

\item{entity_mapping}{The entity to collector mapping created by \code{\link{map_collector}} or manually.}
}
\description{
Parses concentrations from a json trajectory.
}
